# Conflicts:
#	build.gradle
#	common/src/main/java/com/wynntils/commands/BombBellCommand.java
#	common/src/main/java/com/wynntils/commands/CompassCommand.java
#	common/src/main/java/com/wynntils/commands/ServerCommand.java
#	common/src/main/java/com/wynntils/commands/UpdateCommand.java
#	common/src/main/java/com/wynntils/commands/WynntilsCommand.java
#	common/src/main/java/com/wynntils/core/WynntilsMod.java
#	common/src/main/java/com/wynntils/core/chat/tabs/ChatTabModel.java
#	common/src/main/java/com/wynntils/core/features/FeatureRegistry.java
#	common/src/main/java/com/wynntils/features/user/AbbreviateMobHealthFeature.java
#	common/src/main/java/com/wynntils/features/user/ChatItemFeature.java
#	common/src/main/java/com/wynntils/features/user/ChatTimestampFeature.java
#	common/src/main/java/com/wynntils/features/user/CommandsFeature.java
#	common/src/main/java/com/wynntils/features/user/InfoMessageFilterFeature.java
#	common/src/main/java/com/wynntils/features/user/ItemLockFeature.java
#	common/src/main/java/com/wynntils/features/user/ItemScreenshotFeature.java
#	common/src/main/java/com/wynntils/features/user/MountHorseHotkeyFeature.java
#	common/src/main/java/com/wynntils/features/user/TerritoryDefenseMessageFeature.java
#	common/src/main/java/com/wynntils/features/user/UpdatesFeature.java
#	common/src/main/java/com/wynntils/features/user/WorldWaypointDistanceFeature.java
#	common/src/main/java/com/wynntils/features/user/WynncraftPauseScreenFeature.java
#	common/src/main/java/com/wynntils/features/user/inventory/AbilityTreeScrollFeature.java
#	common/src/main/java/com/wynntils/features/user/map/MinimapFeature.java
#	common/src/main/java/com/wynntils/gui/render/RenderUtils.java
#	common/src/main/java/com/wynntils/gui/screens/WynntilsDiscoveriesScreen.java
#	common/src/main/java/com/wynntils/gui/screens/WynntilsQuestBookScreen.java
#	common/src/main/java/com/wynntils/gui/screens/maps/AbstractMapScreen.java
#	common/src/main/java/com/wynntils/gui/screens/maps/GuildMapScreen.java
#	common/src/main/java/com/wynntils/gui/screens/maps/MainMapScreen.java
#	common/src/main/java/com/wynntils/gui/widgets/ChatTabButton.java
#	common/src/main/java/com/wynntils/gui/widgets/DiscoveryButton.java
#	common/src/main/java/com/wynntils/gui/widgets/DiscoveryFilterButton.java
#	common/src/main/java/com/wynntils/gui/widgets/DiscoveryProgressButton.java
#	common/src/main/java/com/wynntils/gui/widgets/TextInputBoxWidget.java
#	common/src/main/java/com/wynntils/mc/EventFactory.java
#	common/src/main/java/com/wynntils/mc/mixin/LocalPlayerMixin.java
#	common/src/main/java/com/wynntils/mc/mixin/ScreenMixin.java
#	common/src/main/java/com/wynntils/wynn/model/ChatItemModel.java
#	common/src/main/java/com/wynntils/wynn/model/discoveries/DiscoveryContainerQueries.java
#	common/src/main/java/com/wynntils/wynn/model/discoveries/DiscoveryManager.java
#	common/src/main/java/com/wynntils/wynn/model/discoveries/objects/DiscoveryInfo.java
#	common/src/main/java/com/wynntils/wynn/model/map/poi/PlayerMainMapPoi.java
#	common/src/main/java/com/wynntils/wynn/model/quests/QuestContainerQueries.java
#	common/src/main/java/com/wynntils/wynn/screens/WynnScreenMatchers.java
#	common/src/main/java/com/wynntils/wynn/utils/WynnItemUtils.java
#	common/src/main/resources/assets/wynntils/lang/en_us.json
#	common/src/main/resources/wynntils.accessWidener
#	common/src/main/resources/wynntils.mixins.json
#	forge/src/main/java/com/wynntils/forge/WynntilsModForge.java
#	forge/src/main/java/com/wynntils/forge/mixins/ForgeGuiMixin.java
#	gradle.properties

+++ common/src/main/java/com/wynntils/gui/render/RenderUtils.java
@@ -227,9 +227,9 @@ public final class RenderUtils {
                 texture.width(),
                 texture.height() / 2f,
                 0,
-                hovered ? texture.height() / 2f : 0,
+                hovered ? texture.height() / 2 : 0,
                 texture.width(),
-                texture.height() / 2f,
+                texture.height() / 2,
                 texture.width(),
                 texture.height());
     }
@@ -253,9 +253,10 @@ public final class RenderUtils {
             float y,
             float width,
             float height,
-            float textureWidth,
-            float textureHeight) {
-        drawTexturedRect(poseStack, tex, x, y, 0, width, height, 0, 0, width, height, textureWidth, textureHeight);
+            int textureWidth,
+            int textureHeight) {
+        drawTexturedRect(
+                poseStack, tex, x, y, 0, width, height, 0, 0, (int) width, (int) height, textureWidth, textureHeight);
     }
 
     public static void drawTexturedRect(
@@ -266,9 +267,10 @@ public final class RenderUtils {
             float z,
             float width,
             float height,
-            float textureWidth,
-            float textureHeight) {
-        drawTexturedRect(poseStack, tex, x, y, z, width, height, 0, 0, width, height, textureWidth, textureHeight);
+            int textureWidth,
+            int textureHeight) {
+        drawTexturedRect(
+                poseStack, tex, x, y, z, width, height, 0, 0, (int) width, (int) height, textureWidth, textureHeight);
     }
 
     public static void drawTexturedRect(
@@ -279,12 +281,12 @@ public final class RenderUtils {
             float z,
             float width,
             float height,
-            float uOffset,
-            float vOffset,
-            float u,
-            float v,
-            float textureWidth,
-            float textureHeight) {
+            int uOffset,
+            int vOffset,
+            int u,
+            int v,
+            int textureWidth,
+            int textureHeight) {
         float uScale = 1f / textureWidth;
         float vScale = 1f / textureHeight;
 
@@ -321,8 +323,8 @@ public final class RenderUtils {
             float z,
             float width,
             float height,
-            float textureWidth,
-            float textureHeight) {
+            int textureWidth,
+            int textureHeight) {
         drawTexturedRect(
                 poseStack, tex, x, y, z, width, height, 0, 0, textureWidth, textureHeight, textureWidth, textureHeight);
     }
@@ -335,10 +337,23 @@ public final class RenderUtils {
             float z,
             float width,
             float height,
-            float textureWidth,
-            float textureHeight) {
+            int textureWidth,
+            int textureHeight) {
         drawTexturedRectWithColor(
-                new PoseStack(), tex, color, x, y, z, width, height, 0, 0, width, height, textureWidth, textureHeight);
+                new PoseStack(),
+                tex,
+                color,
+                x,
+                y,
+                z,
+                width,
+                height,
+                0,
+                0,
+                (int) width,
+                (int) height,
+                textureWidth,
+                textureHeight);
     }
 
     public static void drawTexturedRectWithColor(
@@ -350,12 +365,12 @@ public final class RenderUtils {
             float z,
             float width,
             float height,
-            float uOffset,
-            float vOffset,
-            float u,
-            float v,
-            float textureWidth,
-            float textureHeight) {
+            int uOffset,
+            int vOffset,
+            int u,
+            int v,
+            int textureWidth,
+            int textureHeight) {
         float uScale = 1f / textureWidth;
         float vScale = 1f / textureHeight;
 
@@ -1089,10 +1104,10 @@ public final class RenderUtils {
             float y1,
             float x2,
             float y2,
-            float tx1,
-            float ty1,
-            float tx2,
-            float ty2) {
+            int tx1,
+            int ty1,
+            int tx2,
+            int ty2) {
         // See https://gist.github.com/burgerguy/8233170683ad93eea6aa27ee02a5c4d1
 
         GL11.glEnable(GL11.GL_STENCIL_TEST);
